servers/data-transformation/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "data-mcp-server"
version = "0.1.0"
description = "MCP server for data format transformations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "pandas>=2.2.0",
    "xmltodict>=0.13.0",
    "pyyaml>=6.0",
    "openpyxl>=3.1.0",
    "boto3>=1.34.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["data_mcp_server*"]
